backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.5.3
pydantic-settings==2.1.0
singlestoredb==1.0.0
python-dotenv==1.0.0
python-multipart==0.0.6
websockets==12.0
boto3==1.34.0
botocore==1.34.0

